+++ scripts/events.js
@@ -1,32 +1,32 @@
-carregarTabela();
+// carregarTabela();
 
-// Salvar
-// pegarLocalStorage();
-carregarSelectBatatas()
+// // Salvar
 
-
-// Decorações
-setInterval(estatisticas, 2500)
-
-// Conquistas
-verificarConquistas()
-verificarConquistasInuteis()
-carregarConquistas()
-
-setInterval(() =>{
-  verificarConquistas();
-  verificarConquistasInuteis();
-}, 5000)
-
-// Script
-batatasPorSeg()
-carregarTextosPrecos()
-textoAudio()
-
-setInterval(salvarTemporario, 30000);
-setInterval(batatasPorSeg, 1000);
+// carregarSelectBatatas()
 
 
-document.getElementById("selectBatatasPorSegundo").addEventListener("change", () =>{
-  document.getElementById("estatBatatasPorSegundo").innerHTML = valorSelectCPS()
-});
+// // Decorações
+// // setInterval(estatisticas, 2500)
+
+// // Conquistas
+// verificarConquistas()
+// verificarConquistasInuteis()
+// carregarConquistas()
+
+// // setInterval(() =>{
+// //   verificarConquistas();
+// //   verificarConquistasInuteis();
+// // }, 5000)
+
+// // Script
+// batatasPorSeg()
+// carregarTextosPrecos()
+// textoAudio()
+
+// // setInterval(salvarTemporario, 30000);
+// // setInterval(batatasPorSeg, 1000);
+
+
+// document.getElementById("selectBatatasPorSegundo").addEventListener("change", () =>{
+//   document.getElementById("estatBatatasPorSegundo").innerHTML = valorSelectCPS()
+// });